Mentors comments

+++ db/migrations/000002_create_orders_table.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS orders(
     number BIGINT PRIMARY KEY,
     login  VARCHAR (50) REFERENCES users(login),
     status VARCHAR (50) DEFAULT 'NEW',
-    accrual REAL DEFAULT NULL,
-    withdraw REAL DEFAULT NULL,
+    accrual DECIMAL DEFAULT NULL,
+    withdraw DECIMAL DEFAULT NULL,
     uploaded_at TIMESTAMP DEFAULT now()
 );
